requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-MongoEngine
Flask-JWT-Extended
python-dotenv
